Refactor gamepad module and reorganize module structure.
