+++ crates/bevy_input/src/devices/gamepad/axis/gamepad_axis.rs
@@ -1,0 +1,7 @@
+//! Gamepad input axis implementation module
+use crate::axis_type::GamepadAxisType;
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct GamepadAxis(pub Gamepad, pub GamepadAxisType);
+
