+++ crates/bevy_input/src/devices/gamepad/axis/settings.rs
@@ -1,22 +1,8 @@
+//! Gamepad input axis configuration module
+
 use core::default::Default;
 use core::option::Option::Some;
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-pub enum GamepadAxisType {
-    LeftStickX,
-    LeftStickY,
-    LeftZ,
-    RightStickX,
-    RightStickY,
-    RightZ,
-    DPadX,
-    DPadY,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-pub struct GamepadAxis(pub Gamepad, pub GamepadAxisType);
+use crate::GamepadAxisType;
 
 #[derive(Debug, Clone)]
 pub struct AxisSettings {
@@ -58,14 +44,3 @@ impl AxisSettings {
         new_value
     }
 }
-
-const ALL_AXIS_TYPES: [GamepadAxisType; 8] = [
-    GamepadAxisType::LeftStickX,
-    GamepadAxisType::LeftStickY,
-    GamepadAxisType::LeftZ,
-    GamepadAxisType::RightStickX,
-    GamepadAxisType::RightStickY,
-    GamepadAxisType::RightZ,
-    GamepadAxisType::DPadX,
-    GamepadAxisType::DPadY,
-];
