+++ crates/bevy_input/src/devices/gamepad/button/button_device.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct GamepadButton(pub Gamepad, pub GamepadButtonType);
+
