+++ crates/bevy_input/src/devices/gamepad/button/button_settings.rs
@@ -1,0 +1,28 @@
+use core::default::Default;
+
+#[derive(Debug, Clone)]
+pub struct ButtonSettings {
+    pub press: f32,
+    pub release: f32,
+}
+
+impl Default for ButtonSettings {
+    fn default() -> Self {
+        ButtonSettings {
+            press: 0.75,
+            release: 0.65,
+        }
+    }
+}
+
+impl ButtonSettings {
+    fn is_pressed(&self, value: f32) -> bool {
+        value >= self.press
+    }
+
+    fn is_released(&self, value: f32) -> bool {
+        value <= self.release
+    }
+}
+
+
