+++ crates/bevy_input/src/devices/gamepad/device/gamepad_settings.rs
