+++ crates/bevy_input/src/devices/gamepad/device/mod.rs
@@ -1,0 +1,11 @@
+pub mod gamepad;
+pub mod gamepad_plugin;
+pub mod gamepad_settings;
+
+// Export public exports
+pub use prelude::*;
+
+/// Public exports module
+pub mod prelude {
+    pub use super::{gamepad::Gamepad, gamepad_settings::GamepadSettings, gamepad_plugin::GamepadInputPlugin};
+}
