+++ crates/bevy_input/src/devices/gamepad/event/event_type.rs
@@ -1,0 +1,8 @@
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub enum GamepadEventType {
+    Connected,
+    Disconnected,
+    ButtonChanged(GamepadButtonType, f32),
+    AxisChanged(GamepadAxisType, f32),
+}
