+++ crates/bevy_input/src/devices/gamepad/event/gamepad_event.rs
@@ -1,13 +1,4 @@
-use bevy_input::gamepad::{GamepadButtonType, GamepadAxisType};
-
-#[derive(Debug, Clone, PartialEq)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-pub enum GamepadEventType {
-    Connected,
-    Disconnected,
-    ButtonChanged(GamepadButtonType, f32),
-    AxisChanged(GamepadAxisType, f32),
-}
+use bevy_input::devices::gamepad::event::GamepadEventType;
 
 #[derive(Debug, Clone, PartialEq)]
 #[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
