+++ crates/bevy_input/src/devices/keyboard/input.rs
@@ -1,9 +1,9 @@
 //! Keyboard input implementation module
 
-use crate::{core::ElementState, keyboard::KeyCode};
+use crate::{core::ElementState, keyboard_devices::KeyCode};
 use bevy_app::prelude::*;
 
-/// A key input event from a keyboard device
+/// A key input event from a keyboard_device device
 #[derive(Debug, Clone)]
 pub struct KeyboardInput {
     pub scan_code: u32,
@@ -11,7 +11,7 @@ pub struct KeyboardInput {
     pub state: ElementState,
 }
 
-/// State used by the keyboard input system
+/// State used by the keyboard_device input system
 #[derive(Default)]
 pub struct KeyboardInputState {
     pub(crate) keyboard_input_event_reader: EventReader<KeyboardInput>,
