+++ crates/bevy_input/src/devices/keyboard/mod.rs
@@ -1,6 +1,6 @@
 pub mod input;
-pub mod input_system;
-pub mod keyboard_plugin;
+pub mod system;
+pub mod plugin;
 pub mod keycode;
 
 // Export public exports module
@@ -10,8 +10,8 @@ pub use prelude::*;
 pub mod prelude {
     pub use super::{
         input::{KeyboardInput, KeyboardInputState},
-        input_system::keyboard_input_system,
-        keyboard_plugin::KeyboardInputPlugin,
+        system::keyboard_input_system,
+        plugin::KeyboardInputPlugin,
         keycode::KeyCode,
     };
 }
