+++ crates/bevy_input/src/devices/keyboard/system.rs
@@ -2,7 +2,7 @@
 
 use crate::{
     core::{ElementState, Input},
-    keyboard::prelude::*,
+    keyboard_devices::prelude::*,
 };
 use bevy_app::prelude::*;
 use bevy_ecs::{Local, Res, ResMut};
