+++ crates/bevy_input/src/devices/mod.rs
@@ -1,0 +1,8 @@
+pub mod gamepad;
+pub mod keyboard;
+pub mod mouse;
+pub mod touch;
+
+pub use prelude::*;
+
+pub mod prelude {}
