+++ crates/bevy_input/src/devices/mouse/mod.rs
@@ -3,8 +3,8 @@ use bevy_app::prelude::{EventReader, Events};
 use bevy_ecs::{Local, Res, ResMut};
 use bevy_math::Vec2;
 
-pub mod mouseplugin;
-pub use mouseplugin::MouseInputPlugin;
+pub mod mouse_plugin;
+pub use mouse_plugin::MouseInputPlugin;
 
 /// A mouse button input event
 #[derive(Debug, Clone)]
