--- crates/bevy_input/src/gamepad/gamepad_button.rs
@@ -1,113 +1,0 @@
-use core::default::Default;
-use core::option::Option::Some;
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-pub enum GamepadButtonType {
-    South,
-    East,
-    North,
-    West,
-    C,
-    Z,
-    LeftTrigger,
-    LeftTrigger2,
-    RightTrigger,
-    RightTrigger2,
-    Select,
-    Start,
-    Mode,
-    LeftThumb,
-    RightThumb,
-    DPadUp,
-    DPadDown,
-    DPadLeft,
-    DPadRight,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-pub struct GamepadButton(pub Gamepad, pub GamepadButtonType);
-
-#[derive(Debug, Clone)]
-pub struct ButtonAxisSettings {
-    pub high: f32,
-    pub low: f32,
-    pub threshold: f32,
-}
-
-impl Default for ButtonAxisSettings {
-    fn default() -> Self {
-        ButtonAxisSettings {
-            high: 0.95,
-            low: 0.05,
-            threshold: 0.01,
-        }
-    }
-}
-
-impl ButtonAxisSettings {
-    fn filter(&self, new_value: f32, old_value: Option<f32>) -> f32 {
-        if let Some(old_value) = old_value {
-            if (new_value - old_value).abs() <= self.threshold {
-                return old_value;
-            }
-        }
-        if new_value <= self.low {
-            return 0.0;
-        }
-        if new_value >= self.high {
-            return 1.0;
-        }
-        new_value
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct ButtonSettings {
-    pub press: f32,
-    pub release: f32,
-}
-
-impl Default for ButtonSettings {
-    fn default() -> Self {
-        ButtonSettings {
-            press: 0.75,
-            release: 0.65,
-        }
-    }
-}
-
-impl ButtonSettings {
-    fn is_pressed(&self, value: f32) -> bool {
-        value >= self.press
-    }
-
-    fn is_released(&self, value: f32) -> bool {
-        value <= self.release
-    }
-}
-
-
-const ALL_BUTTON_TYPES: [GamepadButtonType; 19] = [
-    GamepadButtonType::South,
-    GamepadButtonType::East,
-    GamepadButtonType::North,
-    GamepadButtonType::West,
-    GamepadButtonType::C,
-    GamepadButtonType::Z,
-    GamepadButtonType::LeftTrigger,
-    GamepadButtonType::LeftTrigger2,
-    GamepadButtonType::RightTrigger,
-    GamepadButtonType::RightTrigger2,
-    GamepadButtonType::Select,
-    GamepadButtonType::Start,
-    GamepadButtonType::Mode,
-    GamepadButtonType::LeftThumb,
-    GamepadButtonType::RightThumb,
-    GamepadButtonType::DPadUp,
-    GamepadButtonType::DPadDown,
-    GamepadButtonType::DPadLeft,
-    GamepadButtonType::DPadRight,
-];
-
